+++ stm32/cores/maple/libmaple/adafruit_featherlib.c
@@ -48,6 +48,7 @@ extern void USBSerial_callback(uint32_t eid, void* p_data);
 
 //extern void adafruit_wifi_connect_callback(void);
 extern void adafruit_wifi_disconnect_callback(void);
+extern void adafruit_softap_event_callback(uint32_t event, const uint8_t mac[6] );
 
 extern err_t adafruit_tcpserver_connect_callback    (void* socket, void* p_tcpserver);
 extern err_t adafruit_tcp_receive_callback    (void* socket, void* p_tcp);
@@ -72,6 +73,7 @@ ATTR_USED adafruit_arduino_t const adafruit_arduino =
     // wifi
 //    .wifi_connect_callback    = adafruit_wifi_connect_callback,
     .wifi_disconnect_callback   = adafruit_wifi_disconnect_callback,
+    .softap_event_callback      = adafruit_softap_event_callback,
 
     // TCP
     .tcpserver_connect_callback = adafruit_tcpserver_connect_callback,